ParserCommon.h: _ReadStringSCI: fix parsing "\\"
In QFG4CD, script 54 "Import" tries to add a backslash to a string, using the syntax `{\\}`. The parser used a naive test to see if the closing quote char was escpaed, by checking if the previous character was a backslash, which falls over when _that_ backslash is escaped by another backslash. I've fixed it by instead checking the "fEscape" boolean.

I've also added an explicit case for parsing an escaped backslash. This might fix/impact the "default else" case - we may want to uncomment the statement `processCharNormally = true`, since my changes might fix the case for `sciAudio\\command.con`.