Compile.cpp: remove optimization of indexed arrays lookups that use an immediate number value
In QFG4CD, room 54 has `local63`, and array of 34 values. There are two indexed lookups into the array:

[gMaskView [local63 24]]

and

[gMaskView [local63 30]]

Compiling the script would optimize away the indirect lookup from the array, and instead reference the value as a separate local.

Original:

	[local63 34] = [0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 41]

After compiling & decompiling:

	[local63 24] = [0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 16 17 18 19 20 21 22 23 24]
	local140 =  25
	[theGNewStr_3_2 5] = [26 27 28 29 30]
	local142 =  31
	[newStr_4 3] = [32 33 41]

It seems there is a bug in AutoDetectVariableNames: `local142` already exists as an array: `[local142 25] = [...]`, but one of the new locals has also been given the name `local142`.

If we remove the optimization, it avoids this `local142` naming clash. It also helps to keep the compilation "stable": that is, decompiling, then compiling, then decompiling, will give the same decompilation result.

In the future, it would be nice to identify and fix the cause of the bug in AutoDetectVariableNames.

The function `CanDoIndexOptimization` was used as part of error checking as well as optimization, so I've renamed it to `IsIndexedByImmediateNumber`, and it no longer mutates its inputs (i.e. it doesn't perform the optimization).