GitHub Actions: add nightly build/release
Nightly builds run at 7:15 AM UTC. They will create a new build and tag and release. Artifacts are renamed, replacing the SHA with the date. A release is not created if the current commit already has a tag beginning with `nightly/*`.

I've changed the existing build steps into a reusable workflow, and added a separate `on-push.yaml` file.