Fix local procedure not being disassembled/decompiled if it has the same address as an exported (heap) object.
In QFG4CD, script 29 exports two instances, at addresses 4 and 24 (0x18). It also has a local procedure, which _also_ happens to be at address 24 (0x18). I assume it's because the address offsets have a different starting address.

When disassembling/decompiling the script, any identified procedures are removed from `internalProcOffetsTO` if the address is also found in `_exportsTO`.

I found that we can avoid the issue by calling `IsExportAProcedure()` when iterating over `_exportsTO`.

I've tested this by decompiling all the QFG4 scripts, and only script 29 is affected by the change.