Export All Tool - Export as single cell PNG
Exporting will now export to single cells and will have the option to save to a PNG32 with alpha transparencies. It additionally exports the priority image and an experimental export of priority images. This has been cleaned up and optimized, but could use a little work around the interface / dialog box. I will try to get around to working on this a little more, but I thought this was a useful commit to start with.

Also, I fixed a small bug in the .rc file where an #endif was not placed correctly. Originally I wanted to edit the dialog box through visual studio toolbox, but it corrupts anytime I save it. I think the format is different, and or the file has been manually edited so much that the visual editor no longer works. It would be nice to figure this out.